requirements.txt
aiohttp==2.0.7
appdirs==1.4.3
async-timeout==1.2.1
BitstampClient==2.1.0
chardet==3.0.3
configparser==3.5.0
multidict==2.1.5
packaging==16.8
pyparsing==2.2.0
requests==2.14.2
six==1.10.0
telepot==11.0
urllib3==1.21.1
yarl==0.10.2
